new-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
paho-mqtt==1.6.1
PyJWT==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0